+++ src/Cache/Cache.php
@@ -3,45 +3,56 @@
 namespace Sajadsdi\LaravelSettingPro\Cache;
 
 use Sajadsdi\LaravelSettingPro\Contracts\CacheDriverInterface;
-use Illuminate\Support\Facades\Cache as laravelCache;
 
 class Cache implements CacheDriverInterface
 {
-    private array $config;
+    private CacheDriverInterface $driver;
 
     /**
      * @param array $config
      */
     public function __construct(array $config)
     {
-        $this->config = $config;
+        $this->driver = new $config['drivers'][$config['default']]['class']($config['drivers'][$config['default']]);
     }
 
     /**
+     * Get cache from driver with key.
      * @param $key
      * @return mixed
      */
     public function get($key): mixed
     {
-        return laravelCache::get($this->config['prefix'] . '.' . $key);
+        return $this->driver->get($key);
     }
 
     /**
+     * Set cache on driver with key and data.
      * @param string $key
      * @param mixed $data
      * @return void
      */
     public function set(string $key, mixed $data): void
     {
-        laravelCache::put($this->config['prefix'] . '.' . $key, $data);
+        $this->driver->set($key, $data);
     }
 
     /**
+     * Clear cache from driver with key.
      * @param $key
      * @return void
      */
     public function clear($key): void
     {
-        laravelCache::forget($this->config['prefix'] . '.' . $key);
+        $this->driver->clear($key);
+    }
+
+    /**
+     * Clear all caches from driver.
+     * @return void
+     */
+    public function clearAll(): void
+    {
+        $this->driver->clearAll();
     }
 }